Add context update
Updated usecontext based on Neil's help in the slack.
can now pass either the value of darktheme or the current theme (pre-modificated for light/dark moded) into the context value.
the remaining issue at this point is getting the children to recognise that the context is there, regardless of what I put I get "Cannot find name X" error ts2304