added theme object file
Moved the theme object into a separate file to tidy up parent.tsx.
At this point the tutorials i've found are saying to set up the context and provider in the parent then export them (not sure how since they are in a jsx return) and them import them in the children, but that seems to defeat the whole point of using context.