confusing change
I did what the tutorials said and imported the child directly in the childern, the values are now passing but the dark theme toggle has been cut out of the loop and no longer work.

this is very confusing since if we have to manualy import the context in each child why not just import the values?
also this approach cuts out the dark theme toggle since usestate cant be top level, but themecontext has to in order to export.
will revert this, just wanted to post because it confuses me