Finaly functional change context
added a very verbose way of getting the state to change, based on ~5 tutorials, the docs, the recording and the slides.
There is definitely a cleaner way to do this and i'd like to know what it is but this has been driving me round in circles and up the wall for the past 2 days.
set the theme locally as shown in one of the slides, then used a verbose but certain to work set current theme function (trying to update it directly kept running into read-only errors and I couldn't figure out how to just change the 1st 2 values so I replace it with a new value based on.
removed unnececary imports, exports and attributes from themeobject.
updated child 2 to use the context.

The next step would defiantly be to refactor this into something nicer, but I'm saving a working version first :)